Rename modules --> lib
