+++ src/server/api/foods/MakeKrabbyPattyEndpoint.ts
@@ -1,12 +1,12 @@
 import { ApiEndpoint } from "../ApiEndpoint";
-import { Clients } from "src/server/modules";
+import { Clients } from "src/server/lib";
 
 export class MakeKrabbyPattyEndpoint extends ApiEndpoint {
   async handler(req, res) {
     const options = req.body;
 
     // No need to wrap this in a try catch. Let errors bubble up to the global error handler
-    const krabbyPatty = await Clients.krabbyPattyModule.makeKrabbyPatty(options);
+    const krabbyPatty = await Clients.krabbyPattyLib.makeKrabbyPatty(options);
 
     res.json({ krabbyPatty });
   }
