+++ src/server/lib/Clients.ts
@@ -1,9 +1,9 @@
 /**
  * The Clients class is intended to simplify the process of passing dependencies (backend clients
- * and frontend server modules) around the server codebase.
+ * and frontend server libs) around the server codebase.
  *
- * Rather than passing all needed clients and modules to each endpoint handler and module as
- * constructor params, we can load everything onto the Clients class. Endpoint handlers and modules
+ * Rather than passing all needed clients and libs to each endpoint handler and module as
+ * constructor params, we can load everything onto the Clients class. Endpoint handlers and libs
  * can then import this one class for access to all needed dependencies.
  */
 
@@ -11,15 +11,15 @@
 // class.
 // import * as SpongeBobService from "@clever/spongebob-service";
 
-import { KrabbyPattyModule } from "./krabbyPatties/KrabbyPattyModule";
+import { KrabbyPattyLib } from "./krabbyPatties/KrabbyPattyLib";
 
 interface ClientMap {
-  krabbyPattyModule?: any;
+  krabbyPattyLib?: any;
   spongeBobService?: any;
 }
 
 export class Clients {
-  static _krabbyPattyModule;
+  static _krabbyPattyLib;
   static _spongeBobService;
 
   static constructWagClientOptions() {
@@ -27,26 +27,26 @@ export class Clients {
   }
 
   static initialize() {
-    const krabbyPattyModule = new KrabbyPattyModule();
+    const krabbyPattyLib = new KrabbyPattyLib();
 
     // const spongeBobService = new SpongeBobService(this.constructWagClientOptions());
 
     this.loadClients({
-      krabbyPattyModule,
+      krabbyPattyLib,
       // spongeBobService,
     });
   }
 
   static loadClients(clients: ClientMap) {
-    this._krabbyPattyModule = clients.krabbyPattyModule;
+    this._krabbyPattyLib = clients.krabbyPattyLib;
     this._spongeBobService = clients.spongeBobService;
+  }
+
+  static get krabbyPattyLib(): KrabbyPattyLib {
+    return this._krabbyPattyLib;
   }
 
   static get spongeBobService() {
     return this._spongeBobService;
-  }
-
-  static get krabbyPattyModule(): KrabbyPattyModule {
-    return this._krabbyPattyModule;
   }
 }
