--- src/server/modules/krabbyPatties/index.ts
@@ -1,1 +1,0 @@
-export * from "./KrabbyPattyModule";
